menambahkan sourcecode 23-10-2025 11:28 wib

+++ routes/web.php
@@ -2,6 +2,7 @@
 
 use App\Http\Controllers\AuthController;
 use App\Http\Controllers\ResidentController;
+use App\Http\Controllers\UserController;
 use Illuminate\Support\Facades\Route;
 
 // auth
@@ -16,14 +17,17 @@ Route::post('/register',[AuthController::class, 'register']);
 // endauth
 Route::get('/dashboard', function () {
     return view('pages.dashboard');
-})->middleware('role');
+})->middleware('role:Admin,User');
 
-Route::get('/resident',[ResidentController::class, 'index']);
-Route::get('/resident/create',[ResidentController::class, 'create']);
-Route::get('/resident/{id}',[ResidentController::class, 'edit']);
-Route::post('/resident',[ResidentController::class, 'store']);
-Route::put('/resident/{id}',[ResidentController::class, 'update']);
-Route::delete('/resident/{id}',[ResidentController::class, 'destroy']);
+Route::get('/resident',[ResidentController::class, 'index'])->middleware('role:Admin');
+Route::get('/resident/create',[ResidentController::class, 'create'])->middleware('role:Admin');
+Route::get('/resident/{id}',[ResidentController::class, 'edit'])->middleware('role:Admin');
+Route::post('/resident',[ResidentController::class, 'store'])->middleware('role:Admin');
+Route::put('/resident/{id}',[ResidentController::class, 'update'])->middleware('role:Admin');
+Route::delete('/resident/{id}',[ResidentController::class, 'destroy'])->middleware('role:Admin');
+
+// Account-request
+Route::get('/account-request',[UserController::class, 'account_request_view']);
